New plugin type 'local_plugin'
A local plugin doesn't look for assets from the gem path, but instead
operates on the users local files. An example of such a plugin would be
a Sass stylesheet generator. Sure you could release that as a simple
generator for Jekyll, but using ThemeKit, makes it easy to combine the
output with the other stylesheets into a single fingerprinted file and
automatically included in the stylesheet's link tag.

- Themes load before plugins load before local_plugins
- Sass files are now combined with CSS files
- Sass files are removed from Jekyll's static_files array
- All plugins now use register_plugin